+++ packages/visual-dev/src/__tests__/storyshots.test.ts
@@ -3,4 +3,6 @@ import { imageSnapshot } from "@storybook/addon-storyshots-puppeteer";
 
 initStoryshots({
   test: imageSnapshot({ storybookUrl: "http://localhost:9001/" }),
+  storyKindRegex: /^((?!.*?(Loading|Success|loading|success|tooltip|Tooltip)).)*$/,
+  storyNameRegex: /^((?!.*?(Loading|Success|loading|success|tooltip|Tooltip)).)*$/,
 });