+++ packages/visual-dev/src/components/Icon/Icon.stories.tsx
@@ -177,3 +177,7 @@ export const tooltipCustomCSS = () => (
     />
   </TestTooltipDiv>
 );
+export const first = () => <IconView icon={"first_badge"} />;
+export const infinity = () => <IconView icon={"infinity"} />;
+export const gift_filled = () => <IconView icon={"gift_filled"} />;
+export const data = () => <IconView icon={"data"} />;